Stopped Lists from Getting Serialized
Lists attached to gameObjects (e.g. GameBoardSpaceIds) have the
potential to get massively long.  I added the [JsonIgnore] attribute to
all of the Iists I added boolean Contains methods for each list so you
can query the list with the API to see if a list contains a particular
item.  That's all for tonight folks!